modificación del botón eliminar

+++ js/categorias.js
@@ -26,15 +26,27 @@ const { createApp } = Vue
                 })
         },
         eliminar(categoria) {
-            const url = this.url+'/' + categoria;
-            var options = {
-                method: 'DELETE',
-            }
-            fetch(url, options)
-                .then(res => res.text()) // or res.json()
-                .then(res => {
-                    location.reload();
-                })
+            Swal.fire({
+                title: '¿Estás seguro?',
+                text: "¡No podrás revertir esto!",
+                icon: 'Advertencia',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: '¡Sí, elimínalo!'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                    const url = this.url+'/' + categoria;
+                        var options = {
+                            method: 'DELETE',
+                        }
+                        fetch(url, options)
+                            .then(res => res.text()) // or res.json()
+                            .then(res => {
+                                location.reload();
+                        })       
+                }
+              })
         },
         grabar(){
             let categoria = {
